docs/SpiderMonkey/2022-01-12.html

<!DOCTYPE html><html><head>
  <meta charset="UTF-8">
  <title>SpiderMonkey on 2022-01-12</title>
  <link rel="stylesheet" href="../style.css">
  <script>window.room = "SpiderMonkey";</script><script src="../render-roomlist.js"></script>
  <script src="../logs.js"></script>
</head>
<body><div class="wrapper">
<div class="sidebar">
<div class="title">SpiderMonkey<br>2022-01-12<br><a href="plaintext/">plaintext logs</a></div>
<div class="nav">
<a href="2022-01-11" class="nav-link"><span>prev</span></a>
<a href="2022-01-13" class="nav-link"><span style="float:right">next</span></a>
</div>
    <div class="all-rooms"><noscript>JavaScript is required to load the channel index, but you can go to <a href="../..">the static index</a> directly.</noscript></div>
<div class="footer"><a href="https://github.com/bakkot/matrix-archive-bot">source on github</a></div>
</div>
<div class="rhs">
<div class="rhs-header">
<span id="error" style="color: red; display:none">error</span>

<input type="text" id="query" size="25" placeholder="Search SpiderMonkey">
<a id="search-submit" class="button icon-link" title="Search">
  <svg class="icon" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512"><path d="M505 442.7L405.3 343c-4.5-4.5-10.6-7-17-7H372c27.6-35.3 44-79.7 44-128C416 93.1 322.9 0 208 0S0 93.1 0 208s93.1 208 208 208c48.3 0 92.7-16.4 128-44v16.3c0 6.4 2.5 12.5 7 17l99.7 99.7c9.4 9.4 24.6 9.4 33.9 0l28.3-28.3c9.4-9.4 9.4-24.6.1-34zM208 336c-70.7 0-128-57.2-128-128 0-70.7 57.2-128 128-128 70.7 0 128 57.2 128 128 0 70.7-57.2 128-128 128z"></path></svg>
</a>
</div>

<div class="log"><table id="log-table"><tbody id="log-tbody">

  <tr class="msg" id="L0"><td class="ts-cell"><a class="ts" href="#L0" alt="Wed Jan 12 2022 04:08:30 GMT-0800 (Pacific Standard Time)">12:08</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@nbp:mozilla.org">nbp</span>&gt;</div></td><td class="msg-cell">confession: Concurrent delazification is working! :tada: Time to clean-up the diffs and create a few patches to push these changes to the CI and later for review.</td></tr>
  <tr class="msg" id="L1"><td class="ts-cell"><a class="ts" href="#L1" alt="Wed Jan 12 2022 07:15:13 GMT-0800 (Pacific Standard Time)">15:15</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@l11d:mozilla.org">l11d</span>&gt;</div></td><td class="msg-cell">when a LIR instruction performs an oolCallVM and the called native function happens to gc, may this cause the code to be deoptimized?</td></tr>
  <tr class="msg" id="L2"><td class="ts-cell"><a class="ts" href="#L2" alt="Wed Jan 12 2022 07:15:48 GMT-0800 (Pacific Standard Time)">15:15</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title="@jandem:mozilla.org">jandem</span>&gt;</div></td><td class="msg-cell">it's possible yeah</td></tr>
  <tr class="msg" id="L3"><td class="ts-cell"><a class="ts" href="#L3" alt="Wed Jan 12 2022 07:20:40 GMT-0800 (Pacific Standard Time)">15:20</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@l11d:mozilla.org">l11d</span>&gt;</div></td><td class="msg-cell">is it possible to prevent this, at least for testing?</td></tr>
  <tr class="msg" id="L4"><td class="ts-cell"><a class="ts" href="#L4" alt="Wed Jan 12 2022 08:04:04 GMT-0800 (Pacific Standard Time)">16:04</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@nbp:mozilla.org">nbp</span>&gt;</div></td><td class="msg-cell"><span class="nick-4">l11d</span>: callVM basically have the possibility to re-enter JS code, so this would be hard to garantee that we can prevent it in all cases.</td></tr>
  <tr class="msg" id="L5"><td class="ts-cell"><a class="ts" href="#L5" alt="Wed Jan 12 2022 08:04:42 GMT-0800 (Pacific Standard Time)">16:04</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@nbp:mozilla.org">nbp</span>&gt;</div></td><td class="msg-cell">We have a way to prevent GC in limited set of scopes, but this is at your own risk when using it.</td></tr>
  <tr class="msg" id="L6"><td class="ts-cell"><a class="ts" href="#L6" alt="Wed Jan 12 2022 08:05:08 GMT-0800 (Pacific Standard Time)">16:05</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title="@sfink:mozilla.org">sfink</span>&gt;</div></td><td class="msg-cell">I don't think we've ever used it in a scope that can run JS.</td></tr>
  <tr class="msg" id="L7"><td class="ts-cell"><a class="ts" href="#L7" alt="Wed Jan 12 2022 08:06:12 GMT-0800 (Pacific Standard Time)">16:06</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@nbp:mozilla.org">nbp</span>&gt;</div></td><td class="msg-cell">You can find some of these in our GCAPI: <a href="https://searchfox.org/mozilla-central/source/js/public/GCAPI.h#950,975,989">https://searchfox.org/mozilla-central/source/js/public/GCAPI.h#950,975,989</a></td></tr>
  <tr class="msg" id="L8"><td class="ts-cell"><a class="ts" href="#L8" alt="Wed Jan 12 2022 08:07:06 GMT-0800 (Pacific Standard Time)">16:07</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title="@sfink:mozilla.org">sfink</span>&gt;</div></td><td class="msg-cell"><a href="https://searchfox.org/mozilla-central/search?q=symbol:T_js%3A%3Agc%3A%3AAutoSuppressGC&amp;redirect=false">https://searchfox.org/mozilla-central/search?q=symbol:T_js%3A%3Agc%3A%3AAutoSuppressGC&amp;redirect=false</a> gives a good list</td></tr>
  <tr class="msg" id="L9"><td class="ts-cell"><a class="ts" href="#L9" alt="Wed Jan 12 2022 08:15:13 GMT-0800 (Pacific Standard Time)">16:15</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@l11d:mozilla.org">l11d</span>&gt;</div></td><td class="msg-cell">thanks for these pointers; I'm afraid my question was a bit too ambiguous. I'm not looking for a way to prevent the gc but a way to prevent the code from being deoptimized. I suppose this is somehow baked into the engine, at least that's what I understand from this assert: <a href="https://searchfox.org/mozilla-central/source/js/src/gc/Compacting.cpp#413">https://searchfox.org/mozilla-central/source/js/src/gc/Compacting.cpp#413</a> </td></tr>
  <tr class="msg" id="L10"><td class="ts-cell"><a class="ts" href="#L10" alt="Wed Jan 12 2022 08:23:07 GMT-0800 (Pacific Standard Time)">16:23</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@iain:mozilla.org">iain</span>&gt;</div></td><td class="msg-cell"><span class="nick-4">l11d</span>: The problem is that a compacting GC can move objects, and Ion bakes pointers into the code. So it's not safe to do a compacting GC without invalidating Ion code.</td></tr>
  <tr class="msg" id="L11"><td class="ts-cell"><a class="ts" href="#L11" alt="Wed Jan 12 2022 08:24:37 GMT-0800 (Pacific Standard Time)">16:24</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@l11d:mozilla.org">l11d</span>&gt;</div></td><td class="msg-cell">this makes sense. the engine seems to give up after a certain number of recompiles, correct? something like 90?</td></tr>
  <tr class="msg" id="L12"><td class="ts-cell"><a class="ts" href="#L12" alt="Wed Jan 12 2022 08:28:18 GMT-0800 (Pacific Standard Time)">16:28</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@iain:mozilla.org">iain</span>&gt;</div></td><td class="msg-cell">There are some cases where we decide not to recompile afterwards. In non-pathological code, the main reason that would happen is if we are repeatedly bailing out to catch an exception.</td></tr>
  <tr class="msg" id="L13"><td class="ts-cell"><a class="ts" href="#L13" alt="Wed Jan 12 2022 08:29:00 GMT-0800 (Pacific Standard Time)">16:29</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@iain:mozilla.org">iain</span>&gt;</div></td><td class="msg-cell">I believe that will happen if we catch an exception 50 times.</td></tr>
  <tr class="msg" id="L14"><td class="ts-cell"><a class="ts" href="#L14" alt="Wed Jan 12 2022 08:30:22 GMT-0800 (Pacific Standard Time)">16:30</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@nbp:mozilla.org">nbp</span>&gt;</div></td><td class="msg-cell"><span class="nick-4">iain</span>: we do patch pointers which are in Ion.</td></tr>
  <tr class="msg" id="L15"><td class="ts-cell"><a class="ts" href="#L15" alt="Wed Jan 12 2022 08:32:20 GMT-0800 (Pacific Standard Time)">16:32</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@iain:mozilla.org">iain</span>&gt;</div></td><td class="msg-cell"><span class="nick-2">nbp</span>: Oh, yeah, you're right. But I also remember looking into this a little when we were doing Warp, and concluding that there was something we can't update</td></tr>
  <tr class="msg" id="L16"><td class="ts-cell"><a class="ts" href="#L16" alt="Wed Jan 12 2022 08:33:26 GMT-0800 (Pacific Standard Time)">16:33</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@iain:mozilla.org">iain</span>&gt;</div></td><td class="msg-cell"><span class="nick-4">l11d</span>: I'm not aware of a mechanism that would prevent us from recompiling a script based purely on the number of GCs.</td></tr>
  <tr class="msg" id="L17"><td class="ts-cell"><a class="ts" href="#L17" alt="Wed Jan 12 2022 08:39:15 GMT-0800 (Pacific Standard Time)">16:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@l11d:mozilla.org">l11d</span>&gt;</div></td><td class="msg-cell">considering that a GC "counts" as bailing out =&gt; at some point the script will no longer be recompiled, correct?</td></tr>
  <tr class="msg" id="L18"><td class="ts-cell"><a class="ts" href="#L18" alt="Wed Jan 12 2022 08:42:28 GMT-0800 (Pacific Standard Time)">16:42</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@iain:mozilla.org">iain</span>&gt;</div></td><td class="msg-cell"><span class="nick-4">l11d</span>: I would hope not. It would be bad if a long-lived page got slower as we ran GCs and pessimized various scripts. If we run a GC and invalidate scripts that are currently on the stack, then we hit this line during the bailout, which deliberately doesn't do anything to prevent recompilation: <a href="https://searchfox.org/mozilla-central/source/js/src/jit/BaselineBailouts.cpp#2080">https://searchfox.org/mozilla-central/source/js/src/jit/BaselineBailouts.cpp#2080</a></td></tr>
  <tr class="msg" id="L19"><td class="ts-cell"><a class="ts" href="#L19" alt="Wed Jan 12 2022 08:44:17 GMT-0800 (Pacific Standard Time)">16:44</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@iain:mozilla.org">iain</span>&gt;</div></td><td class="msg-cell">I guess in pathological cases we could GC often enough that we don't have time to get back into Ion</td></tr>
  <tr class="msg" id="L20"><td class="ts-cell"><a class="ts" href="#L20" alt="Wed Jan 12 2022 08:48:17 GMT-0800 (Pacific Standard Time)">16:48</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@nbp:mozilla.org">nbp</span>&gt;</div></td><td class="msg-cell"><code>BailoutAction::PreventGC</code> :P</td></tr>
  <tr class="msg" id="L21"><td class="ts-cell"><a class="ts" href="#L21" alt="Wed Jan 12 2022 08:49:28 GMT-0800 (Pacific Standard Time)">16:49</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@iain:mozilla.org">iain</span>&gt;</div></td><td class="msg-cell">It would have to be <code>BailoutAction::RetroactivelyPreventGCWithTimeMachine</code>: we only get there because we've already done the GC and invalidated the script</td></tr>
  <tr class="msg" id="L22"><td class="ts-cell"><a class="ts" href="#L22" alt="Wed Jan 12 2022 08:50:16 GMT-0800 (Pacific Standard Time)">16:50</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@iain:mozilla.org">iain</span>&gt;</div></td><td class="msg-cell">I'm having a hard time remembering why patching pointers isn't sufficient. We cancel off-thread Ion compiles, so it's not untraced off-thread pointers to GC things.</td></tr>
  <tr class="msg" id="L23"><td class="ts-cell"><a class="ts" href="#L23" alt="Wed Jan 12 2022 08:50:53 GMT-0800 (Pacific Standard Time)">16:50</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@iain:mozilla.org">iain</span>&gt;</div></td><td class="msg-cell">Maybe it's possible now?</td></tr>
  <tr class="msg" id="L24"><td class="ts-cell"><a class="ts" href="#L24" alt="Wed Jan 12 2022 08:56:00 GMT-0800 (Pacific Standard Time)">16:56</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@nbp:mozilla.org">nbp</span>&gt;</div></td><td class="msg-cell">My understanding is that we cannot patch locations which are moving nursery pointer to tenured space.</td></tr>
  <tr class="msg" id="L25"><td class="ts-cell"><a class="ts" href="#L25" alt="Wed Jan 12 2022 08:56:37 GMT-0800 (Pacific Standard Time)">16:56</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@nbp:mozilla.org">nbp</span>&gt;</div></td><td class="msg-cell">because of the way to handle barrier (IIRC, do not take that for granted)</td></tr>
  <tr class="msg" id="L26"><td class="ts-cell"><a class="ts" href="#L26" alt="Wed Jan 12 2022 08:57:32 GMT-0800 (Pacific Standard Time)">16:57</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@nbp:mozilla.org">nbp</span>&gt;</div></td><td class="msg-cell">So, if we compile some Ion code with Nursery pointers, we have to invalidate the generated code to recompile with the equivalent tenured pointer.</td></tr>
  <tr class="msg" id="L27"><td class="ts-cell"><a class="ts" href="#L27" alt="Wed Jan 12 2022 08:59:59 GMT-0800 (Pacific Standard Time)">16:59</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title="@wraitii:mozilla.org">wraitii</span>&gt;</div></td><td class="msg-cell">Is there somewhere a list of supported compilers for ESR 91 ? IIRC you now only support clang (and not gcc) ? Which version is the minimum ?</td></tr>
  <tr class="msg" id="L28"><td class="ts-cell"><a class="ts" href="#L28" alt="Wed Jan 12 2022 09:01:18 GMT-0800 (Pacific Standard Time)">17:01</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@nbp:mozilla.org">nbp</span>&gt;</div></td><td class="msg-cell"><span class="nick-4">iain</span>: <a href="https://searchfox.org/mozilla-central/source/js/src/jit/MIROps.yaml#1488-1497">https://searchfox.org/mozilla-central/source/js/src/jit/MIROps.yaml#1488-1497</a></td></tr>
  <tr class="msg" id="L29"><td class="ts-cell"><a class="ts" href="#L29" alt="Wed Jan 12 2022 09:04:10 GMT-0800 (Pacific Standard Time)">17:04</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@iain:mozilla.org">iain</span>&gt;</div></td><td class="msg-cell"><span class="nick-2">nbp</span>: IIRC, that's the solution to nursery pointers that jandem added for Warp: if we want to bake in a nursery pointer, we instead add a level of indirection through a side-table, and update that side-table when we do a minor GC.</td></tr>
  <tr class="msg" id="L30"><td class="ts-cell"><a class="ts" href="#L30" alt="Wed Jan 12 2022 09:04:28 GMT-0800 (Pacific Standard Time)">17:04</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@iain:mozilla.org">iain</span>&gt;</div></td><td class="msg-cell">So it's slightly less efficient, but doesn't require invalidating.</td></tr>
  <tr class="msg" id="L31"><td class="ts-cell"><a class="ts" href="#L31" alt="Wed Jan 12 2022 09:04:51 GMT-0800 (Pacific Standard Time)">17:04</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@nbp:mozilla.org">nbp</span>&gt;</div></td><td class="msg-cell">ok, so that fixes what I recalled being a problem.</td></tr>
  <tr class="msg" id="L32"><td class="ts-cell"><a class="ts" href="#L32" alt="Wed Jan 12 2022 10:03:48 GMT-0800 (Pacific Standard Time)">18:03</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title="@sfink:mozilla.org">sfink</span>&gt;</div></td><td class="msg-cell"><span class="nick-12">wraitii</span>: I'm not sure where to find the officially supported version info, but we do still support both clang and gcc. I know gcc goes back to 9.3.0 since that's what the hazard analysis uses. Looking at treeherder, it looks like it's using clang 12.0.0. Though there are toolchain jobs for clang 5.0.2 and gcc 7.5.0. I'm not sure what those are about.</td></tr>
  <tr class="msg" id="L33"><td class="ts-cell"><a class="ts" href="#L33" alt="Wed Jan 12 2022 10:03:54 GMT-0800 (Pacific Standard Time)">18:03</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title="@sfink:mozilla.org">sfink</span>&gt;</div></td><td class="msg-cell">maybe a question for <a href="https://matrix.to/#/#build:mozilla.org">#build:mozilla.org</a></td></tr>
  <tr class="msg" id="L34"><td class="ts-cell"><a class="ts" href="#L34" alt="Wed Jan 12 2022 10:40:53 GMT-0800 (Pacific Standard Time)">18:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">Here's a question, and a "here's how you'd look for a solution" recommendation is fine.</td></tr>
  <tr class="msg" id="L35"><td class="ts-cell"><a class="ts" href="#L35" alt="Wed Jan 12 2022 10:43:12 GMT-0800 (Pacific Standard Time)">18:43</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">I create an instance of the interpreter, execcute a script, and it segfaults in js::jit::ReprotectRegion when I call JS_DestroyContext on the context the interpreter is using.</td></tr>
  <tr class="msg" id="L36"><td class="ts-cell"><a class="ts" href="#L36" alt="Wed Jan 12 2022 10:46:17 GMT-0800 (Pacific Standard Time)">18:46</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">what's the local variables and parameters there?</td></tr>
  <tr class="msg" id="L37"><td class="ts-cell"><a class="ts" href="#L37" alt="Wed Jan 12 2022 10:47:09 GMT-0800 (Pacific Standard Time)">18:47</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">and what line or what operation crashes?</td></tr>
  <tr class="msg" id="L38"><td class="ts-cell"><a class="ts" href="#L38" alt="Wed Jan 12 2022 10:48:12 GMT-0800 (Pacific Standard Time)">18:48</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@nbp:mozilla.org">nbp</span>&gt;</div></td><td class="msg-cell"><span class="nick-4">chaircrusher</span>: Socket process? Do you have the right to map pages as executable?</td></tr>
  <tr class="msg" id="L39"><td class="ts-cell"><a class="ts" href="#L39" alt="Wed Jan 12 2022 10:48:30 GMT-0800 (Pacific Standard Time)">18:48</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@nbp:mozilla.org">nbp</span>&gt;</div></td><td class="msg-cell">^ cc <span class="nick-10">jandem</span></td></tr>
  <tr class="msg" id="L40"><td class="ts-cell"><a class="ts" href="#L40" alt="Wed Jan 12 2022 10:49:08 GMT-0800 (Pacific Standard Time)">18:49</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@iain:mozilla.org">iain</span>&gt;</div></td><td class="msg-cell">If it's happening during JS_DestroyContext, it's presumably happening when we poison the jitcode here: <a href="https://searchfox.org/mozilla-central/source/js/src/vm/Runtime.cpp#570">https://searchfox.org/mozilla-central/source/js/src/vm/Runtime.cpp#570</a></td></tr>
  <tr class="msg" id="L41"><td class="ts-cell"><a class="ts" href="#L41" alt="Wed Jan 12 2022 10:49:22 GMT-0800 (Pacific Standard Time)">18:49</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">that's right</td></tr>
  <tr class="msg" id="L42"><td class="ts-cell"><a class="ts" href="#L42" alt="Wed Jan 12 2022 10:49:50 GMT-0800 (Pacific Standard Time)">18:49</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@iain:mozilla.org">iain</span>&gt;</div></td><td class="msg-cell">So I would expect that you've already successfully mapped pages as executable</td></tr>
  <tr class="msg" id="L43"><td class="ts-cell"><a class="ts" href="#L43" alt="Wed Jan 12 2022 10:50:00 GMT-0800 (Pacific Standard Time)">18:50</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@iain:mozilla.org">iain</span>&gt;</div></td><td class="msg-cell">And executed code in those pages</td></tr>
  <tr class="msg" id="L44"><td class="ts-cell"><a class="ts" href="#L44" alt="Wed Jan 12 2022 10:50:08 GMT-0800 (Pacific Standard Time)">18:50</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@iain:mozilla.org">iain</span>&gt;</div></td><td class="msg-cell">And it's something about the teardown process going wrong</td></tr>
  <tr class="msg" id="L45"><td class="ts-cell"><a class="ts" href="#L45" alt="Wed Jan 12 2022 10:50:44 GMT-0800 (Pacific Standard Time)">18:50</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">it's in js::jit::ReprotectRegion, last line (in GDB) but I think it must be barfing on an assert or something</td></tr>
  <tr class="msg" id="L46"><td class="ts-cell"><a class="ts" href="#L46" alt="Wed Jan 12 2022 10:51:00 GMT-0800 (Pacific Standard Time)">18:51</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@nbp:mozilla.org">nbp</span>&gt;</div></td><td class="msg-cell"><span class="nick-4">iain</span>: it is possible that the policy to forbid mapping pages as executable got setup after starting the JitRuntime?</td></tr>
  <tr class="msg" id="L47"><td class="ts-cell"><a class="ts" href="#L47" alt="Wed Jan 12 2022 10:51:04 GMT-0800 (Pacific Standard Time)">18:51</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">In my version ProcessExecutableMemory.cpp:730</td></tr>
  <tr class="msg" id="L48"><td class="ts-cell"><a class="ts" href="#L48" alt="Wed Jan 12 2022 10:51:20 GMT-0800 (Pacific Standard Time)">18:51</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">that's ESR78</td></tr>
  <tr class="msg" id="L49"><td class="ts-cell"><a class="ts" href="#L49" alt="Wed Jan 12 2022 10:51:38 GMT-0800 (Pacific Standard Time)">18:51</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell"><a href="https://searchfox.org/mozilla-esr78/rev/3c633b1a0994f380032a616eced632398354d83e/js/src/jit/ProcessExecutableMemory.cpp#730">https://searchfox.org/mozilla-esr78/rev/3c633b1a0994f380032a616eced632398354d83e/js/src/jit/ProcessExecutableMemory.cpp#730</a> this line?  or different line?</td></tr>
  <tr class="msg" id="L50"><td class="ts-cell"><a class="ts" href="#L50" alt="Wed Jan 12 2022 10:52:19 GMT-0800 (Pacific Standard Time)">18:52</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">yup</td></tr>
  <tr class="msg" id="L51"><td class="ts-cell"><a class="ts" href="#L51" alt="Wed Jan 12 2022 10:52:58 GMT-0800 (Pacific Standard Time)">18:52</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">but 730 is the closing brace for ReprotectRegion and it's compiled optimized so the function parameters are optimized out.</td></tr>
  <tr class="msg" id="L52"><td class="ts-cell"><a class="ts" href="#L52" alt="Wed Jan 12 2022 10:53:10 GMT-0800 (Pacific Standard Time)">18:53</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">can you get disassembly for the function? it might tell more about what the operation is</td></tr>
  <tr class="msg" id="L53"><td class="ts-cell"><a class="ts" href="#L53" alt="Wed Jan 12 2022 10:53:52 GMT-0800 (Pacific Standard Time)">18:53</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@nbp:mozilla.org">nbp</span>&gt;</div></td><td class="msg-cell"><span class="nick-4">chaircrusher</span>: Linux / Mac / Windows?</td></tr>
  <tr class="msg" id="L54"><td class="ts-cell"><a class="ts" href="#L54" alt="Wed Jan 12 2022 10:54:04 GMT-0800 (Pacific Standard Time)">18:54</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">CENTOS8</td></tr>
  <tr class="msg" id="L55"><td class="ts-cell"><a class="ts" href="#L55" alt="Wed Jan 12 2022 10:54:31 GMT-0800 (Pacific Standard Time)">18:54</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">(might be nice to check if the same crash happens with debug+non-optimized build</td></tr>
  <tr class="msg" id="L56"><td class="ts-cell"><a class="ts" href="#L56" alt="Wed Jan 12 2022 10:54:41 GMT-0800 (Pacific Standard Time)">18:54</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@nbp:mozilla.org">nbp</span>&gt;</div></td><td class="msg-cell">ok, Linux properly handles OOM on mprotect. (as opposed to all others)</td></tr>
  <tr class="msg" id="L57"><td class="ts-cell"><a class="ts" href="#L57" alt="Wed Jan 12 2022 10:56:23 GMT-0800 (Pacific Standard Time)">18:56</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">debug build might also catch something before reaching the crash</td></tr>
  <tr class="msg" id="L58"><td class="ts-cell"><a class="ts" href="#L58" alt="Wed Jan 12 2022 10:59:12 GMT-0800 (Pacific Standard Time)">18:59</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">anyone have any good guides, or pointers to good mochitests that use iframes to test cross compartment wrappers </td></tr>
  <tr class="msg" id="L59"><td class="ts-cell"><a class="ts" href="#L59" alt="Wed Jan 12 2022 10:59:38 GMT-0800 (Pacific Standard Time)">18:59</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">rebuild in progress</td></tr>
  <tr class="msg" id="L60"><td class="ts-cell"><a class="ts" href="#L60" alt="Wed Jan 12 2022 10:59:40 GMT-0800 (Pacific Standard Time)">18:59</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">I am apparently bad at web stuff, and so I keep getting either access denied issues, or references that don't seem to be CCWs, </td></tr>
  <tr class="msg" id="L61"><td class="ts-cell"><a class="ts" href="#L61" alt="Wed Jan 12 2022 11:01:02 GMT-0800 (Pacific Standard Time)">19:01</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">(I want to write a mochitest that involves readable streams with the source from another page, and what happens when the source disappears; which i'm pretty sure is possible, but how this looks on the web is apparently still mysterious to me) </td></tr>
  <tr class="msg" id="L62"><td class="ts-cell"><a class="ts" href="#L62" alt="Wed Jan 12 2022 11:01:16 GMT-0800 (Pacific Standard Time)">19:01</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">(be back in an hour so; lunch time here) </td></tr>
  <tr class="msg" id="L63"><td class="ts-cell"><a class="ts" href="#L63" alt="Wed Jan 12 2022 11:03:20 GMT-0800 (Pacific Standard Time)">19:03</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@nbp:mozilla.org">nbp</span>&gt;</div></td><td class="msg-cell"><span class="nick-2">mgaudet</span>: I used iframe in the past, but for checking the XDR encoding status, if this helps.</td></tr>
  <tr class="msg" id="L64"><td class="ts-cell"><a class="ts" href="#L64" alt="Wed Jan 12 2022 11:03:25 GMT-0800 (Pacific Standard Time)">19:03</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">I thought iframe with same domain works for that case.  haven't thought much about when it becomes CCW tho</td></tr>
  <tr class="msg" id="L65"><td class="ts-cell"><a class="ts" href="#L65" alt="Wed Jan 12 2022 11:03:36 GMT-0800 (Pacific Standard Time)">19:03</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">isn't it "always" ?</td></tr>
  <tr class="msg" id="L66"><td class="ts-cell"><a class="ts" href="#L66" alt="Wed Jan 12 2022 11:05:44 GMT-0800 (Pacific Standard Time)">19:05</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@nbp:mozilla.org">nbp</span>&gt;</div></td><td class="msg-cell"><span class="nick-2">mgaudet</span>: <a href="https://searchfox.org/mozilla-central/source/dom/base/test/test_script_loader_js_cache.html#51-52">https://searchfox.org/mozilla-central/source/dom/base/test/test_script_loader_js_cache.html#51-52</a> I know iframes have a <code>sandbox</code> attribute or similar which allow faking some of the properties which are expected. Maybe this could be used for testing CCW.</td></tr>
  <tr class="msg" id="L67"><td class="ts-cell"><a class="ts" href="#L67" alt="Wed Jan 12 2022 11:10:28 GMT-0800 (Pacific Standard Time)">19:10</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@evilpie:mozilla.org">evilpie</span>&gt;</div></td><td class="msg-cell">The sandbox attribute is not related to CCW, I think it might actually make some CCW impossible</td></tr>
  <tr class="msg" id="L68"><td class="ts-cell"><a class="ts" href="#L68" alt="Wed Jan 12 2022 11:10:44 GMT-0800 (Pacific Standard Time)">19:10</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">oh, iframe and toplevel are in the same compartment with simple case</td></tr>
  <tr class="msg" id="L69"><td class="ts-cell"><a class="ts" href="#L69" alt="Wed Jan 12 2022 11:11:41 GMT-0800 (Pacific Standard Time)">19:11</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">btw, what's the purpose of the iframe test?  is it with promise job + dying global?  if so, CCW isn't necessary but you just need to remove iframe from the DOM tree</td></tr>
  <tr class="msg" id="L70"><td class="ts-cell"><a class="ts" href="#L70" alt="Wed Jan 12 2022 11:19:20 GMT-0800 (Pacific Standard Time)">19:19</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">looks like, you can use <code><a href="http://window.open">window.open</a></code> to create 2 compartments accessible each other</td></tr>
  <tr class="msg" id="L71"><td class="ts-cell"><a class="ts" href="#L71" alt="Wed Jan 12 2022 11:24:15 GMT-0800 (Pacific Standard Time)">19:24</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">Well I did recompile SpiderMonkey ESR78 with debug on and now I'm getting assertion failures in JS_Init</td></tr>
  <tr class="msg" id="L72"><td class="ts-cell"><a class="ts" href="#L72" alt="Wed Jan 12 2022 11:25:49 GMT-0800 (Pacific Standard Time)">19:25</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">Assertion failure: !aOther.IsNull() (Cannot compute with aOther null value), at /home/develop/pagewiz/build/js-build/dist/include/mozilla/TimeStamp.h:557
</td></tr>
  <tr class="msg" id="L73"><td class="ts-cell"><a class="ts" href="#L73" alt="Wed Jan 12 2022 11:26:04 GMT-0800 (Pacific Standard Time)">19:26</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">it doesn't fail if I compile SpiderMonkey optimized.</td></tr>
  <tr class="msg" id="L74"><td class="ts-cell"><a class="ts" href="#L74" alt="Wed Jan 12 2022 11:26:15 GMT-0800 (Pacific Standard Time)">19:26</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">Is there something I'm supposed to do BEFORE calling JS_Init?</td></tr>
  <tr class="msg" id="L75"><td class="ts-cell"><a class="ts" href="#L75" alt="Wed Jan 12 2022 11:26:20 GMT-0800 (Pacific Standard Time)">19:26</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">what's the backtrace?</td></tr>
  <tr class="msg" id="L76"><td class="ts-cell"><a class="ts" href="#L76" alt="Wed Jan 12 2022 11:28:33 GMT-0800 (Pacific Standard Time)">19:28</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell"><pre><code class="language-#0">#1  0x0000000001ccbb22 in mozilla::TimeStamp::ProcessCreation (aIsInconsistent=0x0) at /home/develop/pagewiz/build/js/mozglue/misc/TimeStamp.cpp:91
#2  0x0000000000ac2e7a in JS::detail::InitWithFailureDiagnostic (isDebugBuild=true) at /home/develop/pagewiz/build/js/js/src/vm/Initialization.cpp:142
#3  0x0000000000719159 in JS_Init () at /develop/pagewiz/build/js-build/dist/include/js/Initialization.h:63
#4  0x00000000007048c8 in ScriptData::ScriptData (this=0x4590a20) at /develop/pagewiz/pagewiz/ScriptEngine.cpp:610
#5  0x0000000000705196 in ScriptEngine::ScriptEngine (this=0x454cd30 &lt;_ZL8jsEngine&gt;) at /develop/pagewiz/pagewiz/ScriptEngine.cpp:705
#6  0x00000000007896ad in __static_initialization_and_destruction_0 (__initialize_p=1, __priority=65535) at /develop/pagewiz/pagewiz/TSCompile.cpp:68
#7  0x00000000007896d7 in _GLOBAL__sub_I_FromTSCompiler () at /develop/pagewiz/pagewiz/TSCompile.cpp:159
#8  0x000000000250746d in __libc_csu_init ()
#9  0x00007ffff260b41e in __libc_start_main () from /lib64/libc.so.6
#10 0x00000000006eb04e in _start ()
</code></pre>
</td></tr>
  <tr class="msg" id="L77"><td class="ts-cell"><a class="ts" href="#L77" alt="Wed Jan 12 2022 11:31:05 GMT-0800 (Pacific Standard Time)">19:31</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">how are you supposed to bracket included text? <code>like this</code></td></tr>
  <tr class="msg" id="L78"><td class="ts-cell"><a class="ts" href="#L78" alt="Wed Jan 12 2022 11:31:12 GMT-0800 (Pacific Standard Time)">19:31</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">or...</td></tr>
  <tr class="msg" id="L79"><td class="ts-cell"><a class="ts" href="#L79" alt="Wed Jan 12 2022 11:31:34 GMT-0800 (Pacific Standard Time)">19:31</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">triple backticks?</td></tr>
  <tr class="msg" id="L80"><td class="ts-cell"><a class="ts" href="#L80" alt="Wed Jan 12 2022 11:32:04 GMT-0800 (Pacific Standard Time)">19:32</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">`wikkid</td></tr>
  <tr class="msg" id="L81"><td class="ts-cell"><a class="ts" href="#L81" alt="Wed Jan 12 2022 11:33:21 GMT-0800 (Pacific Standard Time)">19:33</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">I cannot think of the case that hits the assertion.  I think it's <code>mozilla::TimeStampInitialization::mFirstTimeStamp</code> field, that gets initialized in constructor</td></tr>
  <tr class="msg" id="L82"><td class="ts-cell"><a class="ts" href="#L82" alt="Wed Jan 12 2022 11:33:51 GMT-0800 (Pacific Standard Time)">19:33</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell"><a href="https://searchfox.org/mozilla-esr78/rev/3c633b1a0994f380032a616eced632398354d83e/mozglue/misc/TimeStamp.cpp#38">https://searchfox.org/mozilla-esr78/rev/3c633b1a0994f380032a616eced632398354d83e/mozglue/misc/TimeStamp.cpp#38</a></td></tr>
  <tr class="msg" id="L83"><td class="ts-cell"><a class="ts" href="#L83" alt="Wed Jan 12 2022 11:34:46 GMT-0800 (Pacific Standard Time)">19:34</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@evilpie:mozilla.org">evilpie</span>&gt;</div></td><td class="msg-cell">Are you calling JS_Init from a static constructor?</td></tr>
  <tr class="msg" id="L84"><td class="ts-cell"><a class="ts" href="#L84" alt="Wed Jan 12 2022 11:34:54 GMT-0800 (Pacific Standard Time)">19:34</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@evilpie:mozilla.org">evilpie</span>&gt;</div></td><td class="msg-cell">maybe that is causing the issues</td></tr>
  <tr class="msg" id="L85"><td class="ts-cell"><a class="ts" href="#L85" alt="Wed Jan 12 2022 11:34:57 GMT-0800 (Pacific Standard Time)">19:34</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">oh</td></tr>
  <tr class="msg" id="L86"><td class="ts-cell"><a class="ts" href="#L86" alt="Wed Jan 12 2022 11:34:58 GMT-0800 (Pacific Standard Time)">19:34</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">it looks like sInitOnce.timestamp is zeros</td></tr>
  <tr class="msg" id="L87"><td class="ts-cell"><a class="ts" href="#L87" alt="Wed Jan 12 2022 11:35:01 GMT-0800 (Pacific Standard Time)">19:35</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">no</td></tr>
  <tr class="msg" id="L88"><td class="ts-cell"><a class="ts" href="#L88" alt="Wed Jan 12 2022 11:35:27 GMT-0800 (Pacific Standard Time)">19:35</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell"><code>__static_initialization_and_destruction_</code> in the backtrace</td></tr>
  <tr class="msg" id="L89"><td class="ts-cell"><a class="ts" href="#L89" alt="Wed Jan 12 2022 11:36:02 GMT-0800 (Pacific Standard Time)">19:36</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">so, it's constructed before <code>sInitOnce</code></td></tr>
  <tr class="msg" id="L90"><td class="ts-cell"><a class="ts" href="#L90" alt="Wed Jan 12 2022 11:36:15 GMT-0800 (Pacific Standard Time)">19:36</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">That's called before the first time I create a context and global object.</td></tr>
  <tr class="msg" id="L91"><td class="ts-cell"><a class="ts" href="#L91" alt="Wed Jan 12 2022 11:36:33 GMT-0800 (Pacific Standard Time)">19:36</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">ok so no static initialization of interpreters is what you're telling me.</td></tr>
  <tr class="msg" id="L92"><td class="ts-cell"><a class="ts" href="#L92" alt="Wed Jan 12 2022 11:36:52 GMT-0800 (Pacific Standard Time)">19:36</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">because of static initialization in SpiderMonkey</td></tr>
  <tr class="msg" id="L93"><td class="ts-cell"><a class="ts" href="#L93" alt="Wed Jan 12 2022 11:37:23 GMT-0800 (Pacific Standard Time)">19:37</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">yeah, you cannot call JSAPI from static constructor</td></tr>
  <tr class="msg" id="L94"><td class="ts-cell"><a class="ts" href="#L94" alt="Wed Jan 12 2022 11:37:43 GMT-0800 (Pacific Standard Time)">19:37</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">Is that written down somewhere I forgot to read?</td></tr>
  <tr class="msg" id="L95"><td class="ts-cell"><a class="ts" href="#L95" alt="Wed Jan 12 2022 11:39:01 GMT-0800 (Pacific Standard Time)">19:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">uh, I'm not aware of any doc, but in general, it's better avoid calling any API within static constructor that you cannot guarantee the order</td></tr>
  <tr class="msg" id="L96"><td class="ts-cell"><a class="ts" href="#L96" alt="Wed Jan 12 2022 11:39:03 GMT-0800 (Pacific Standard Time)">19:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@evilpie:mozilla.org">evilpie</span>&gt;</div></td><td class="msg-cell">I don't think that has occurred to anyone before, it's certainly impossible in Gecko</td></tr>
  <tr class="msg" id="L97"><td class="ts-cell"><a class="ts" href="#L97" alt="Wed Jan 12 2022 11:40:46 GMT-0800 (Pacific Standard Time)">19:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">anyway, I think that could be the cause of the above crash, that the internal state is broken due to the unexpected order of initialization</td></tr>
  <tr class="msg" id="L98"><td class="ts-cell"><a class="ts" href="#L98" alt="Wed Jan 12 2022 11:48:09 GMT-0800 (Pacific Standard Time)">19:48</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">You were correct sir!</td></tr>
  <tr class="msg" id="L99"><td class="ts-cell"><a class="ts" href="#L99" alt="Wed Jan 12 2022 11:48:46 GMT-0800 (Pacific Standard Time)">19:48</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">but that just moves me to my next assertion failure: <code>Assertion failure: mStatementDone (Guard object should not be used as a temporary.), at /develop/pagewiz/build/js-build/dist/include/mozilla/GuardObjects.h:11</code></td></tr>
  <tr class="msg" id="L100"><td class="ts-cell"><a class="ts" href="#L100" alt="Wed Jan 12 2022 11:51:48 GMT-0800 (Pacific Standard Time)">19:51</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell"><p>This is where I'm going to call JS_Execute Script.</p>
<pre><code class="language-JS::RootedValue">			JSAutoRealm(context,code);
			bool successfulExecute = JS_ExecuteScript(context, code, &amp;result);
</code></pre>
</td></tr>
  <tr class="msg" id="L101"><td class="ts-cell"><a class="ts" href="#L101" alt="Wed Jan 12 2022 11:52:37 GMT-0800 (Pacific Standard Time)">19:52</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@iain:mozilla.org">iain</span>&gt;</div></td><td class="msg-cell">Comment on <a href="https://searchfox.org/mozilla-esr78/source/mfbt/GuardObjects.h#109-111">that assertion</a></td></tr>
  <tr class="msg" id="L102"><td class="ts-cell"><a class="ts" href="#L102" alt="Wed Jan 12 2022 11:53:25 GMT-0800 (Pacific Standard Time)">19:53</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@iain:mozilla.org">iain</span>&gt;</div></td><td class="msg-cell"><pre><code>     * Assert that the guard object was not used as a temporary.  (Note that
     * this assert might also fire if init is not called because the guard
     * object's implementation is not using the above macros correctly.)
</code></pre>
</td></tr>
  <tr class="msg" id="L103"><td class="ts-cell"><a class="ts" href="#L103" alt="Wed Jan 12 2022 11:55:11 GMT-0800 (Pacific Standard Time)">19:55</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@iain:mozilla.org">iain</span>&gt;</div></td><td class="msg-cell">Oh, but you're calling it via JSAutoRealm, which presumably should use the macros properly. Huh.</td></tr>
  <tr class="msg" id="L104"><td class="ts-cell"><a class="ts" href="#L104" alt="Wed Jan 12 2022 11:55:26 GMT-0800 (Pacific Standard Time)">19:55</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">weird</td></tr>
  <tr class="msg" id="L105"><td class="ts-cell"><a class="ts" href="#L105" alt="Wed Jan 12 2022 11:55:41 GMT-0800 (Pacific Standard Time)">19:55</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">looking at the ESR78 js shell to see what it does</td></tr>
  <tr class="msg" id="L106"><td class="ts-cell"><a class="ts" href="#L106" alt="Wed Jan 12 2022 11:55:53 GMT-0800 (Pacific Standard Time)">19:55</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell"><code>JSAuto*</code> must be allocated on stack</td></tr>
  <tr class="msg" id="L107"><td class="ts-cell"><a class="ts" href="#L107" alt="Wed Jan 12 2022 11:56:05 GMT-0800 (Pacific Standard Time)">19:56</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">so, <code>JSAutoRealm ar(context, code);</code></td></tr>
  <tr class="msg" id="L108"><td class="ts-cell"><a class="ts" href="#L108" alt="Wed Jan 12 2022 11:57:01 GMT-0800 (Pacific Standard Time)">19:57</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@iain:mozilla.org">iain</span>&gt;</div></td><td class="msg-cell">Oh, good eye</td></tr>
  <tr class="msg" id="L109"><td class="ts-cell"><a class="ts" href="#L109" alt="Wed Jan 12 2022 11:57:31 GMT-0800 (Pacific Standard Time)">19:57</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">the code above discards the instance immediately, and your remaining code doesn't enter the realm</td></tr>
  <tr class="msg" id="L110"><td class="ts-cell"><a class="ts" href="#L110" alt="Wed Jan 12 2022 11:58:14 GMT-0800 (Pacific Standard Time)">19:58</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">shit that's a typo and very weird, isn't it?</td></tr>
  <tr class="msg" id="L111"><td class="ts-cell"><a class="ts" href="#L111" alt="Wed Jan 12 2022 11:59:58 GMT-0800 (Pacific Standard Time)">19:59</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">Moral of the story I was doing stuff I shouldn't because I never compiled ESR78 in debug mode.</td></tr>
  <tr class="msg" id="L112"><td class="ts-cell"><a class="ts" href="#L112" alt="Wed Jan 12 2022 12:02:28 GMT-0800 (Pacific Standard Time)">20:02</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell"><p>OK finally I get down to it my original problem (crash on JS_DestroyContext:</p>
<pre><code class="language-Hit"></code></pre>
</td></tr>
  <tr class="msg" id="L113"><td class="ts-cell"><a class="ts" href="#L113" alt="Wed Jan 12 2022 12:03:11 GMT-0800 (Pacific Standard Time)">20:03</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">Hit MOZ_CRASH(There should not be any roots during runtime shutdown) at /home/develop/pagewiz/build/js/js/src/gc/RootMarking.cpp:455</td></tr>
  <tr class="msg" id="L114"><td class="ts-cell"><a class="ts" href="#L114" alt="Wed Jan 12 2022 12:03:53 GMT-0800 (Pacific Standard Time)">20:03</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell"><mx-reply><blockquote> btw, what's the purpose of the iframe test?  is it with promise job + dying global?  if so, CCW isn't necessary but you just need to remove iframe from the DOM tree</blockquote></mx-reply>I'd like to cover that case too -- so that's good to hear; but right now I'm sort of interested in what happens when the source object becomes a DeadObjectProxy</td></tr>
  <tr class="msg" id="L115"><td class="ts-cell"><a class="ts" href="#L115" alt="Wed Jan 12 2022 12:05:13 GMT-0800 (Pacific Standard Time)">20:05</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">I've got a test written using nukeSandbox, but I'd sort of like to see what other browsers do, so I'm trying to figure out first what happens for us in a similar case. </td></tr>
  <tr class="msg" id="L116"><td class="ts-cell"><a class="ts" href="#L116" alt="Wed Jan 12 2022 12:09:01 GMT-0800 (Pacific Standard Time)">20:09</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell"><span class="nick-4">chaircrusher</span>: if you have <code>JS::Rooted</code> on the stack, you must leave the scope before calling <code>JS_DestroyContext</code></td></tr>
  <tr class="msg" id="L117"><td class="ts-cell"><a class="ts" href="#L117" alt="Wed Jan 12 2022 12:09:44 GMT-0800 (Pacific Standard Time)">20:09</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">I suspect it's because I keep a global around in a persistentrooted</td></tr>
  <tr class="msg" id="L118"><td class="ts-cell"><a class="ts" href="#L118" alt="Wed Jan 12 2022 12:09:53 GMT-0800 (Pacific Standard Time)">20:09</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title="@sfink:mozilla.org">sfink</span>&gt;</div></td><td class="msg-cell">that would do it</td></tr>
  <tr class="msg" id="L119"><td class="ts-cell"><a class="ts" href="#L119" alt="Wed Jan 12 2022 12:09:59 GMT-0800 (Pacific Standard Time)">20:09</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">There are no rooted on the stack at this point iirc</td></tr>
  <tr class="msg" id="L120"><td class="ts-cell"><a class="ts" href="#L120" alt="Wed Jan 12 2022 12:10:49 GMT-0800 (Pacific Standard Time)">20:10</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title="@sfink:mozilla.org">sfink</span>&gt;</div></td><td class="msg-cell">reset the PersistentRooted before destroying</td></tr>
  <tr class="msg" id="L121"><td class="ts-cell"><a class="ts" href="#L121" alt="Wed Jan 12 2022 12:11:17 GMT-0800 (Pacific Standard Time)">20:11</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title="@sfink:mozilla.org">sfink</span>&gt;</div></td><td class="msg-cell"><code>reset()</code> removes from the root lists</td></tr>
  <tr class="msg" id="L122"><td class="ts-cell"><a class="ts" href="#L122" alt="Wed Jan 12 2022 12:11:39 GMT-0800 (Pacific Standard Time)">20:11</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title="@sfink:mozilla.org">sfink</span>&gt;</div></td><td class="msg-cell">(as opposed to just emptying out the stored value)</td></tr>
  <tr class="msg" id="L123"><td class="ts-cell"><a class="ts" href="#L123" alt="Wed Jan 12 2022 12:16:28 GMT-0800 (Pacific Standard Time)">20:16</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">yup</td></tr>
  <tr class="msg" id="L124"><td class="ts-cell"><a class="ts" href="#L124" alt="Wed Jan 12 2022 12:20:25 GMT-0800 (Pacific Standard Time)">20:20</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell"><span class="nick-2">mgaudet</span>: then I think you can use the <code><a href="http://window.open">window.open</a></code> with same domain</td></tr>
  <tr class="msg" id="L125"><td class="ts-cell"><a class="ts" href="#L125" alt="Wed Jan 12 2022 12:21:37 GMT-0800 (Pacific Standard Time)">20:21</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">not sure how to generate <code>DeadObjectProxy</code> tho.  maybe you can use privileged API or something?</td></tr>
  <tr class="msg" id="L126"><td class="ts-cell"><a class="ts" href="#L126" alt="Wed Jan 12 2022 12:22:13 GMT-0800 (Pacific Standard Time)">20:22</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">just closing the opened window doesn't do that</td></tr>
  <tr class="msg" id="L127"><td class="ts-cell"><a class="ts" href="#L127" alt="Wed Jan 12 2022 12:22:55 GMT-0800 (Pacific Standard Time)">20:22</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">Hrm. Ok. </td></tr>
  <tr class="msg" id="L128"><td class="ts-cell"><a class="ts" href="#L128" alt="Wed Jan 12 2022 12:22:58 GMT-0800 (Pacific Standard Time)">20:22</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell"><pre><code>var w = window.open("other.html");
w.addEventListener("load", () =&gt; {
  var foo = w.foo;
  w.close();
  ...
});
</code></pre>
</td></tr>
  <tr class="msg" id="L129"><td class="ts-cell"><a class="ts" href="#L129" alt="Wed Jan 12 2022 12:24:27 GMT-0800 (Pacific Standard Time)">20:24</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">(this will also hit the dying global case tho</td></tr>
  <tr class="msg" id="L130"><td class="ts-cell"><a class="ts" href="#L130" alt="Wed Jan 12 2022 12:26:52 GMT-0800 (Pacific Standard Time)">20:26</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">Ok. This is more complicated than I thought! My mental model seems to be wrong, as I'd assumed that any time you had a reference into another iframe, you'd get a CCW which would then be nuked when you navigated the iframe the old page contents would be lost, and all wrappers replaced with DeadObjectProxies </td></tr>
  <tr class="msg" id="L131"><td class="ts-cell"><a class="ts" href="#L131" alt="Wed Jan 12 2022 12:31:07 GMT-0800 (Pacific Standard Time)">20:31</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">I guess something around this?  <a href="https://searchfox.org/mozilla-central/rev/5ce782034b736293836a2850ab698e149ca04657/dom/base/WindowDestroyedEvent.cpp#101-143">https://searchfox.org/mozilla-central/rev/5ce782034b736293836a2850ab698e149ca04657/dom/base/WindowDestroyedEvent.cpp#101-143</a>  I don't know much about the code around that</td></tr>
  <tr class="msg" id="L132"><td class="ts-cell"><a class="ts" href="#L132" alt="Wed Jan 12 2022 12:31:17 GMT-0800 (Pacific Standard Time)">20:31</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">effing gdb</td></tr>
  <tr class="msg" id="L133"><td class="ts-cell"><a class="ts" href="#L133" alt="Wed Jan 12 2022 12:32:02 GMT-0800 (Pacific Standard Time)">20:32</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">every so often when debugging issues in SpiderMonkey gdb gets hung up in a thread and I have to break it out and continue to keep going</td></tr>
  <tr class="msg" id="L134"><td class="ts-cell"><a class="ts" href="#L134" alt="Wed Jan 12 2022 12:48:43 GMT-0800 (Pacific Standard Time)">20:48</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">Well resetting the only persistentRooted in my app still barfs in JS_DestroyContext</td></tr>
  <tr class="msg" id="L135"><td class="ts-cell"><a class="ts" href="#L135" alt="Wed Jan 12 2022 12:49:07 GMT-0800 (Pacific Standard Time)">20:49</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">same assertion?</td></tr>
  <tr class="msg" id="L136"><td class="ts-cell"><a class="ts" href="#L136" alt="Wed Jan 12 2022 12:49:16 GMT-0800 (Pacific Standard Time)">20:49</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">yeah</td></tr>
  <tr class="msg" id="L137"><td class="ts-cell"><a class="ts" href="#L137" alt="Wed Jan 12 2022 12:49:23 GMT-0800 (Pacific Standard Time)">20:49</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">trying to track it down</td></tr>
  <tr class="msg" id="L138"><td class="ts-cell"><a class="ts" href="#L138" alt="Wed Jan 12 2022 12:54:52 GMT-0800 (Pacific Standard Time)">20:54</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title="@chaircrusher:mozilla.org">chaircrusher</span>&gt;</div></td><td class="msg-cell">just figured out something not to do: don't call init on a PersisitentRooted twice </td></tr>
  <tr class="msg" id="L139"><td class="ts-cell"><a class="ts" href="#L139" alt="Wed Jan 12 2022 12:55:23 GMT-0800 (Pacific Standard Time)">20:55</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title="@sfink:mozilla.org">sfink</span>&gt;</div></td><td class="msg-cell">in the past, I've stepped through <code>GCRuntime::traceRuntimeCommon</code> (might be a little different in your version), stepping over each statement and seeing which one pushes to the mark stack.</td></tr>
  <tr class="msg" id="L140"><td class="ts-cell"><a class="ts" href="#L140" alt="Wed Jan 12 2022 12:56:04 GMT-0800 (Pacific Standard Time)">20:56</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title="@sfink:mozilla.org">sfink</span>&gt;</div></td><td class="msg-cell">(maybe by setting a hw watchpoint on <code>&amp;((GCMarker*)trc)-&gt;currentStackPtr.value</code>, in which case you can let it continue until it hits)</td></tr>
  <tr class="msg" id="L141"><td class="ts-cell"><a class="ts" href="#L141" alt="Wed Jan 12 2022 13:00:10 GMT-0800 (Pacific Standard Time)">21:00</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title="@sfink:mozilla.org">sfink</span>&gt;</div></td><td class="msg-cell">actually, that would just show when you switch stacks. Maybe put a breakpoint in GCMarker::pushTaggedPtr (by line number)?</td></tr>
  <tr class="msg" id="L142"><td class="ts-cell"><a class="ts" href="#L142" alt="Wed Jan 12 2022 13:19:52 GMT-0800 (Pacific Standard Time)">21:19</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell"><p>confession: Discovering yet new ways JavaScript syntax can surprise me:</p>
<pre><code>js&gt; var cursed = {async "foo.bar"() { print("what") } }
js&gt; Object.getOwnPropertyNames(cursed)                  
["foo.bar"]
</code></pre>
</td></tr>
  <tr class="msg" id="L143"><td class="ts-cell"><a class="ts" href="#L143" alt="Wed Jan 12 2022 13:19:54 GMT-0800 (Pacific Standard Time)">21:19</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title="@botzilla:mozilla.org">botzilla</span>&gt;</div></td><td class="msg-cell">Seen! Your update will eventually appear on <a href="https://robotzilla.github.io/histoire">https://robotzilla.github.io/histoire</a></td></tr>
  <tr class="msg" id="L144"><td class="ts-cell"><a class="ts" href="#L144" alt="Wed Jan 12 2022 14:04:15 GMT-0800 (Pacific Standard Time)">22:04</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">this testcase creates dead wrapper <a href="https://searchfox.org/mozilla-central/source/js/xpconnect/tests/browser/browser_dead_object.js">https://searchfox.org/mozilla-central/source/js/xpconnect/tests/browser/browser_dead_object.js</a></td></tr>
  <tr class="msg" id="L145"><td class="ts-cell"><a class="ts" href="#L145" alt="Wed Jan 12 2022 14:04:28 GMT-0800 (Pacific Standard Time)">22:04</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">but I cannot replicate the same thing with content priv code</td></tr>
  <tr class="msg" id="L146"><td class="ts-cell"><a class="ts" href="#L146" alt="Wed Jan 12 2022 14:07:11 GMT-0800 (Pacific Standard Time)">22:07</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">so I guess it happens only for CCW between chrome and content?</td></tr>
  <tr class="msg" id="L147"><td class="ts-cell"><a class="ts" href="#L147" alt="Wed Jan 12 2022 14:07:41 GMT-0800 (Pacific Standard Time)">22:07</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell"><a href="https://searchfox.org/mozilla-central/rev/5ce782034b736293836a2850ab698e149ca04657/dom/base/WindowDestroyedEvent.cpp#134">https://searchfox.org/mozilla-central/rev/5ce782034b736293836a2850ab698e149ca04657/dom/base/WindowDestroyedEvent.cpp#134</a></td></tr>
  <tr class="msg" id="L148"><td class="ts-cell"><a class="ts" href="#L148" alt="Wed Jan 12 2022 14:10:57 GMT-0800 (Pacific Standard Time)">22:10</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">hrm. </td></tr>
  <tr class="msg" id="L149"><td class="ts-cell"><a class="ts" href="#L149" alt="Wed Jan 12 2022 14:11:10 GMT-0800 (Pacific Standard Time)">22:11</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">That would sort of be fantastic, as it means my testing requirements drop</td></tr>
  <tr class="msg" id="L150"><td class="ts-cell"><a class="ts" href="#L150" alt="Wed Jan 12 2022 14:12:04 GMT-0800 (Pacific Standard Time)">22:12</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">but I was reading this <a href="https://mdn-archive.mossop.dev/en-US/docs/Mozilla/Gecko/Script_security">https://mdn-archive.mossop.dev/en-US/docs/Mozilla/Gecko/Script_security</a> and while it is archived, it fairly clearly says I should expect a wrapper between same origin parges... do we just never clean up a same-origin global if wrappers are still live?</td></tr>
  <tr class="msg" id="L151"><td class="ts-cell"><a class="ts" href="#L151" alt="Wed Jan 12 2022 14:12:32 GMT-0800 (Pacific Standard Time)">22:12</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">so the existence of the wrapper actually prevents the cleanup, so we should never see a dead object proxy in content? </td></tr>
  <tr class="msg" id="L152"><td class="ts-cell"><a class="ts" href="#L152" alt="Wed Jan 12 2022 14:13:19 GMT-0800 (Pacific Standard Time)">22:13</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">yeah, I confirmed that I get CCW with <code><a href="http://window.open">window.open</a></code> case, but that CCW is not nuked on navigation or <code>window.close</code></td></tr>
  <tr class="msg" id="L153"><td class="ts-cell"><a class="ts" href="#L153" alt="Wed Jan 12 2022 14:13:33 GMT-0800 (Pacific Standard Time)">22:13</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">Ok! That's something for sure! </td></tr>
  <tr class="msg" id="L154"><td class="ts-cell"><a class="ts" href="#L154" alt="Wed Jan 12 2022 14:13:38 GMT-0800 (Pacific Standard Time)">22:13</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">and the reference keeps working</td></tr>
  <tr class="msg" id="L155"><td class="ts-cell"><a class="ts" href="#L155" alt="Wed Jan 12 2022 14:14:03 GMT-0800 (Pacific Standard Time)">22:14</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">MDN for JSMSG_DEAD actually says it's only for addons: <a href="https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Errors/Dead_object">https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Errors/Dead_object</a></td></tr>
  <tr class="msg" id="L156"><td class="ts-cell"><a class="ts" href="#L156" alt="Wed Jan 12 2022 14:14:13 GMT-0800 (Pacific Standard Time)">22:14</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">which is also good news</td></tr>
  <tr class="msg" id="L157"><td class="ts-cell"><a class="ts" href="#L157" alt="Wed Jan 12 2022 14:15:05 GMT-0800 (Pacific Standard Time)">22:15</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">Well. Actually no; I guess an Add-on could, hypothetically, create a readable stream from a content object and use it as a source... which is the test case I was trying to write.</td></tr>
  <tr class="msg" id="L158"><td class="ts-cell"><a class="ts" href="#L158" alt="Wed Jan 12 2022 14:16:42 GMT-0800 (Pacific Standard Time)">22:16</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">yeah, add-on, or chrome-priv code</td></tr>
  <tr class="msg" id="L159"><td class="ts-cell"><a class="ts" href="#L159" alt="Wed Jan 12 2022 14:16:53 GMT-0800 (Pacific Standard Time)">22:16</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">like the above dead object testcase</td></tr>
  <tr class="msg" id="L160"><td class="ts-cell"><a class="ts" href="#L160" alt="Wed Jan 12 2022 14:23:09 GMT-0800 (Pacific Standard Time)">22:23</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">Ok. Sweet. I will have to figure out how all that works, but it helps that I'm in a better place now. 

I still have no idea how cross global referencing works but progress. Thanks arai! 

you tested <a href="http://window.open">window.open</a> -- what did that test case look like? I'm really confused by how all this is supposed to work so it keeps going wrong for me</td></tr>
  <tr class="msg" id="L161"><td class="ts-cell"><a class="ts" href="#L161" alt="Wed Jan 12 2022 14:28:04 GMT-0800 (Pacific Standard Time)">22:28</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell"><p>what I was using is something like:</p>
<pre><code>var w = window.open("other.html");
w.addEventListener("load", () =&gt; {
  var foo = w.foo;
  w.close();
  setInterval(() =&gt; {
    foo();
  }, 1000);
});
</code></pre>
</td></tr>
  <tr class="msg" id="L162"><td class="ts-cell"><a class="ts" href="#L162" alt="Wed Jan 12 2022 14:28:46 GMT-0800 (Pacific Standard Time)">22:28</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">with call for modified built-in to print whether a value is CCW or not</td></tr>
  <tr class="msg" id="L163"><td class="ts-cell"><a class="ts" href="#L163" alt="Wed Jan 12 2022 14:29:11 GMT-0800 (Pacific Standard Time)">22:29</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">(not much specific to Promise/Stream)</td></tr>
  <tr class="msg" id="L164"><td class="ts-cell"><a class="ts" href="#L164" alt="Wed Jan 12 2022 14:29:47 GMT-0800 (Pacific Standard Time)">22:29</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">OK; thank you. I'll continue playing </td></tr>
  <tr class="msg" id="L165"><td class="ts-cell"><a class="ts" href="#L165" alt="Wed Jan 12 2022 14:35:48 GMT-0800 (Pacific Standard Time)">22:35</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">oh, if you meant an automated testcase, I wasn't using it. I was using standalone HTML file hosted on local HTTP server, with above script, and manually operating on it</td></tr>
  <tr class="msg" id="L166"><td class="ts-cell"><a class="ts" href="#L166" alt="Wed Jan 12 2022 14:39:08 GMT-0800 (Pacific Standard Time)">22:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">ah. Fair enough. 
</td></tr>
  <tr class="msg" id="L167"><td class="ts-cell"><a class="ts" href="#L167" alt="Wed Jan 12 2022 14:41:31 GMT-0800 (Pacific Standard Time)">22:41</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell"><p>so, in other.html, how do you define 'foo'? I've got</p>
<pre><code>&lt;html&gt;
    &lt;body&gt;
        &lt;script&gt;
        function hi() { console.log("uhm"); return 10; }
        &lt;/script&gt;
    &lt;/body&gt;
&lt;/html&gt;
</code></pre>
<p>and my test case body looks like this:</p>
<pre><code>    SimpleTest.waitForExplicitFinish()

    var w = window.open("test_iframe_helper.html");
    w.addEventListener("load", () =&gt; {
    var hi = w.hi;
    info(hi)
    w.close();
    setInterval(() =&gt; {
        is(hi(), 10);

        SimpleTest.finish()
    }, 1000);
    });
</code></pre>
</td></tr>
  <tr class="msg" id="L168"><td class="ts-cell"><a class="ts" href="#L168" alt="Wed Jan 12 2022 14:41:43 GMT-0800 (Pacific Standard Time)">22:41</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">but in my case, w.hi is coming out as undefined </td></tr>
  <tr class="msg" id="L169"><td class="ts-cell"><a class="ts" href="#L169" alt="Wed Jan 12 2022 14:42:38 GMT-0800 (Pacific Standard Time)">22:42</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">is the opener chrome-priv code?</td></tr>
  <tr class="msg" id="L170"><td class="ts-cell"><a class="ts" href="#L170" alt="Wed Jan 12 2022 14:42:58 GMT-0800 (Pacific Standard Time)">22:42</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">if so, I guess there's Xray interacting</td></tr>
  <tr class="msg" id="L171"><td class="ts-cell"><a class="ts" href="#L171" alt="Wed Jan 12 2022 14:43:57 GMT-0800 (Pacific Standard Time)">22:43</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell"> I have no idea... this is just a mochitest (are mochitests a lot more special than I thought?)</td></tr>
  <tr class="msg" id="L172"><td class="ts-cell"><a class="ts" href="#L172" alt="Wed Jan 12 2022 14:44:31 GMT-0800 (Pacific Standard Time)">22:44</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell"><a href="https://paste.mozilla.org/HGdNE9bR">https://paste.mozilla.org/HGdNE9bR</a></td></tr>
  <tr class="msg" id="L173"><td class="ts-cell"><a class="ts" href="#L173" alt="Wed Jan 12 2022 14:45:01 GMT-0800 (Pacific Standard Time)">22:45</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">and I think you can do <code>var hi = w.eval("(function f() { ... })");</code> in the opener's side to get the actual function object, with CCW?</td></tr>
  <tr class="msg" id="L174"><td class="ts-cell"><a class="ts" href="#L174" alt="Wed Jan 12 2022 14:50:57 GMT-0800 (Pacific Standard Time)">22:50</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell"><span class="nick-2">mgaudet</span>: is it mochitest plain?  I've copy-pasted the same thing and I get <code>hi</code> as function with your code</td></tr>
  <tr class="msg" id="L175"><td class="ts-cell"><a class="ts" href="#L175" alt="Wed Jan 12 2022 14:51:14 GMT-0800 (Pacific Standard Time)">22:51</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">err, no, slightly different</td></tr>
  <tr class="msg" id="L176"><td class="ts-cell"><a class="ts" href="#L176" alt="Wed Jan 12 2022 14:51:32 GMT-0800 (Pacific Standard Time)">22:51</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">How would I tell if plain or not? This was me just trying to create a new set of mochitests in dom/streams</td></tr>
  <tr class="msg" id="L177"><td class="ts-cell"><a class="ts" href="#L177" alt="Wed Jan 12 2022 14:51:48 GMT-0800 (Pacific Standard Time)">22:51</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell"><p>so I have a <code>mochitest.ini</code> that looks like</p>
<pre><code>[DEFAULT]
support-files =
  test_iframe_helper.html

[test_iframes.html]
</code></pre>
</td></tr>
  <tr class="msg" id="L178"><td class="ts-cell"><a class="ts" href="#L178" alt="Wed Jan 12 2022 14:52:00 GMT-0800 (Pacific Standard Time)">22:52</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell"><code>mochitest.ini</code> = plain</td></tr>
  <tr class="msg" id="L179"><td class="ts-cell"><a class="ts" href="#L179" alt="Wed Jan 12 2022 14:52:33 GMT-0800 (Pacific Standard Time)">22:52</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell"><code>browser.ini</code> = mochitest browser, <code>chrome.ini</code> = mochitest browser chrome</td></tr>
  <tr class="msg" id="L180"><td class="ts-cell"><a class="ts" href="#L180" alt="Wed Jan 12 2022 14:53:18 GMT-0800 (Pacific Standard Time)">22:53</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">So... plain? I guess? </td></tr>
  <tr class="msg" id="L181"><td class="ts-cell"><a class="ts" href="#L181" alt="Wed Jan 12 2022 14:53:22 GMT-0800 (Pacific Standard Time)">22:53</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">yes</td></tr>
  <tr class="msg" id="L182"><td class="ts-cell"><a class="ts" href="#L182" alt="Wed Jan 12 2022 14:54:16 GMT-0800 (Pacific Standard Time)">22:54</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">copy-pasted the exactly same thing and still getting right function</td></tr>
  <tr class="msg" id="L183"><td class="ts-cell"><a class="ts" href="#L183" alt="Wed Jan 12 2022 14:55:42 GMT-0800 (Pacific Standard Time)">22:55</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">I'm testing in <code>js/xpconnect/tests/mochitest/</code>, but I guess not much difference?</td></tr>
  <tr class="msg" id="L184"><td class="ts-cell"><a class="ts" href="#L184" alt="Wed Jan 12 2022 14:56:00 GMT-0800 (Pacific Standard Time)">22:56</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">Allllright. i am at end of day, so I will try figuring this out more tomorrow</td></tr>
  <tr class="msg" id="L185"><td class="ts-cell"><a class="ts" href="#L185" alt="Wed Jan 12 2022 14:56:51 GMT-0800 (Pacific Standard Time)">22:56</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title="@arai:mozilla.org">arai</span>&gt;</div></td><td class="msg-cell">okay.  feel free to ping me whenever</td></tr>
  <tr class="msg" id="L186"><td class="ts-cell"><a class="ts" href="#L186" alt="Wed Jan 12 2022 14:57:00 GMT-0800 (Pacific Standard Time)">22:57</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">confession: Spent a lot of the day trying to write tests for things that don't happen -- but now I know what can happen, so hopefully I can write tests for things that <em>do</em> happen tomorrow</td></tr>
  <tr class="msg" id="L187"><td class="ts-cell"><a class="ts" href="#L187" alt="Wed Jan 12 2022 14:57:02 GMT-0800 (Pacific Standard Time)">22:57</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title="@botzilla:mozilla.org">botzilla</span>&gt;</div></td><td class="msg-cell">Seen! Your update will eventually appear on <a href="https://robotzilla.github.io/histoire">https://robotzilla.github.io/histoire</a></td></tr>
  <tr class="msg" id="L188"><td class="ts-cell"><a class="ts" href="#L188" alt="Wed Jan 12 2022 14:58:31 GMT-0800 (Pacific Standard Time)">22:58</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title="@mgaudet:mozilla.org">mgaudet</span>&gt;</div></td><td class="msg-cell">(Thanks for the help Arai; you've been great) </td></tr>

</tbody></table></div></div></div>
</body></html>
